Added Blueprint an restructured app for it
