--- urls.json
@@ -1,1 +1,0 @@
-{"go": {"url": "https://google.com"}, "house": {"file": "househousehouse.jpg"}}
