+++ urlshort/templates/page_not_found.html
@@ -1,0 +1,17 @@
+{% extends 'base.html' %}
+
+{% block title %}Page Not Found{% endblock %}
+
+{% block main %}
+
+
+<div class="container text-center">
+  <h1 class="jumbotron-heading">Page Not Found</h1>
+  <p class="lead">We couldn't find what you are looking for. Come vist our homepage :)</p>
+  <p>
+    <a href="{{ url_for('urlshort.home') }}"><button class="btn btn-primary my-2">Home</button></a>
+  </p>
+</div>
+
+
+{% endblock %}
